Match-id-f3b5fe432e09f6da3c8f862f30857a65580639f0

+++ src/framework/communicator/impl/independent_op/comm_mem_manager.h
@@ -37,7 +37,7 @@ public:
     ~CommMemMgr() = default;
 
      // cclbuffer内存
-    void CommSetHcclBufferSize(uint32_t size);
+    void CommSetHcclBufferSize(uint64_t size);
     HcclResult GetHcclBuffer(CommBuffer *buffer);
     
     // 用户注册/反注册内存
